+++ packages/TSFCore/src/interfaces/Core/TSFCore_apply_op_helper_decl.hpp
@@ -48,9 +48,9 @@ template<class Scalar>
 void apply_op_validate_input(
 	const char                      func_name[]
 	,const RTOpPack::RTOpT<Scalar>  &op
-	,const size_t                   num_vecs
+	,const int                   num_vecs
 	,const Vector<Scalar>*          vecs[]
-	,const size_t                   num_targ_vecs
+	,const int                   num_targ_vecs
 	,Vector<Scalar>*                targ_vecs[]
 	,RTOpPack::ReductTarget         *reduct_obj
 	,const Index                    first_ele
@@ -70,9 +70,9 @@ void apply_op_validate_input(
 template<class Scalar>
 void apply_op_serial(
 	const RTOpPack::RTOpT<Scalar>  &op
-	,const size_t                  num_vecs
+	,const int                  num_vecs
 	,const Vector<Scalar>*         vecs[]
-	,const size_t                  num_targ_vecs
+	,const int                  num_targ_vecs
 	,Vector<Scalar>*               targ_vecs[]
 	,RTOpPack::ReductTarget        *reduct_obj
 	,const Index                   first_ele
